backend/requirements.txt

Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests
Flask-Cors==4.0.0





